[REFACTOR]: splash loading component 추가
기존 App.tsx에서 splash screen 띄우고 로딩하던 것 AppLoading 컴포넌트로 분리
